Adds nested TargetField Capability rescoring and deleting
Previously the ByField Rerank processor was used to handle a field that was only within the first level of the source map, instead customers may have a requirement to have a nested object in their document. This commit adds functionality to be able to handle these scenarios as well as deleting empty objects that are a result of deleting the targetField. Three unit tests were onboarding to ensure that the scores are retrieved and that the search response was able to rerank sucessfully; manual testing was done to see that removal empty maps were done when removing the targetField

Signed-off-by: Brian Flores <iflorbri@amazon.com>