Adds IT for ByFieldRerank (Search Pipline workflow)
This commit adds a IT where the ByFieldRerank processor interacts with the search pipeline workflow. It creates an index with sample data and the runs a query using the Rerank processor. Lastly it observes the result and tests that is has worked by checking the source map and the updated score

Signed-off-by: Brian Flores <iflorbri@amazon.com>