Adds UTs that involve throwing Exceptions for ByField Rerank
In this commit, UTs were added to assert that exceptions were being thrown the way expected (cases include not having a source map to perform reranking or having invalid mappings). There was a section of code that was not able to get hit but couldn't be avoided because the signature of the rescoreSearchResponse that could not throw an exception.

Signed-off-by: Brian Flores <iflorbri@amazon.com>