Adds keep_previous_score field to ByField processor
After offline discussion it was noted that the current rerank implementation does not aggregate another field, thus it was preferred to keep this functionality optional for users who are more savy. Now Users will see a search response with the same _source mapping and updated _score as the default behavior. Additional tests were (UTs) added and (UTs & ITs) modified to catch this new behavior.

Signed-off-by: Brian Flores <iflorbri@amazon.com>