Refactors code to be more readable
The delete target_field with empty maps with method was recursive and needed to be optimized to be executed iteratively. There were also pieces of code that were ambigious as such refactoring was done to simplify the serialization of the sourceAsMap to _source. Tests were rerun to check that the refactor had no effect on the funcitonality

Signed-off-by: Brian Flores <iflorbri@amazon.com>