Extract generic methods in ByFieldRerank to utility class processorSearchResponseUtil
This commit extracts a lot generic methods that involve retrieving data from the source map as well as deleting maps from the targetField. Additionally a functional Interface was introduced that can be used by future rerank processors to be able to validate that their SearchResponse is in correct form to be able to perform reranking. UTs and ITs all passed sucessfully to check that the functionality is in tact

Signed-off-by: Brian Flores <iflorbri@amazon.com>